Extract base application
