+++ packages/cli/source/application.ts
@@ -1,46 +1,12 @@
 import { Identifiers } from "@mainsail/constants";
-import { Container } from "@mainsail/container";
+import { Application as BaseApplication} from "@mainsail/container";
 import type { Contracts } from "@mainsail/contracts";
 import { resolve } from "path";
 
-export class Application {
-	#container: Contracts.Kernel.Container.Container;
-
+export class Application extends BaseApplication {
 	public constructor() {
-		this.#container = new Container();
-		this.#container.bind(Identifiers.Cli.Application.Instance).toConstantValue(this);
-	}
-
-	public bind<T>(
-		serviceIdentifier: Contracts.Kernel.Container.ServiceIdentifier<T>,
-	): Contracts.Kernel.Container.BindToFluentSyntax<T> {
-		return this.#container.bind(serviceIdentifier);
-	}
-
-	public rebind<T>(
-		serviceIdentifier: Contracts.Kernel.Container.ServiceIdentifier<T>,
-	): Contracts.Kernel.Container.BindToFluentSyntax<T> {
-		if (this.#container.isBound(serviceIdentifier)) {
-			this.#container.unbindSync(serviceIdentifier);
-		}
-
-		return this.#container.bind(serviceIdentifier);
-	}
-
-	public unbind<T>(serviceIdentifier: Contracts.Kernel.Container.ServiceIdentifier<T>): void {
-		return this.#container.unbindSync(serviceIdentifier);
-	}
-
-	public get<T>(serviceIdentifier: Contracts.Kernel.Container.ServiceIdentifier<T>): T {
-		return this.#container.get(serviceIdentifier);
-	}
-
-	public isBound<T>(serviceIdentifier: Contracts.Kernel.Container.ServiceIdentifier<T>): boolean {
-		return this.#container.isBound(serviceIdentifier);
-	}
-
-	public resolve<T>(constructorFunction: Contracts.Kernel.Container.Newable<T>): T {
-		return this.#container.get(constructorFunction, { autobind: true });
+		super();
+		this.bind(Identifiers.Cli.Application.Instance).toConstantValue(this);
 	}
 
 	public getCorePath(type: string, file?: string): string {
