+++ packages/container/source/index.ts
@@ -1,2 +1,3 @@
+export * from "./application.js";
 export * from "./decorators.js";
 export * from "./ioc.js";
